# modified laplacian_matrix to achieve speed-up