For some reason sparse + Diagonal is very slow. Modified laplacian_matrix to use sparse + sparse instead. This achieves a 100x speed-up in some cases.