@sbromberger 2019-04-01
Could you post some benchmarks?