@sbromberger 2019-04-03
These benchmarks are really compelling. Thanks.